+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # VisiData version history
 
+# next
+
+- rename skiplines option to skip
+- rename headerlines option to header
+- move Path and subclasses out of vdtui
+- TextSheet source is any iterable of strings
+- Sheet.filetype provides default save filename extension
+
 ## 0.97 (2017-10-05)
 - Features
     - [replay]